split imagenet script for applications
